Restrict Context to apkBundle
The Context interface was converted into an abstract class in order for
the ApkBundle instance to be available in the case of up-casting,
the subclasses and usage of the constructors have been adapted accordingly.

The javadocs could reflect the above explanation. Consider whether it is worth
adding why the apkBundle instance is passed around.